simplificación del codigo en componentes
